+++ packages/sdk-router/src/rfq/fastBridgeRouter.ts
@@ -59,10 +59,10 @@ export class FastBridgeRouter implements SynapseModule {
   private async hydrateCache() {
     if (HYDRATION_SUPPORTED_CHAIN_IDS.includes(this.chainId)) {
       try {
-        await Promise.all([
-          this.getProtocolFeeRate(),
-        ])
-      } catch (e) {}
+        await Promise.all([this.getProtocolFeeRate()])
+      } catch (e) {
+        console.error('fastBridgeRouter: Error hydrating cache', e)
+      }
     }
   }