+++ packages/sdk-router/src/router/synapseCCTPRouter.ts
@@ -56,10 +56,17 @@ export class SynapseCCTPRouter extends Router {
   private async hydrateCache() {
     if (HYDRATION_SUPPORTED_CHAIN_IDS.includes(this.chainId)) {
       try {
+<<<<<<< HEAD
         await Promise.all([
           this.chainGasAmount(),
         ])
       } catch(e) {}
+=======
+        await Promise.all([this.chainGasAmount()])
+      } catch (e) {
+        console.error('synapseCCTPRouter: Error hydrating cache', e)
+      }
+>>>>>>> master
     }
   }