+++ packages/synapse-interface/components/StateManagedBridge/BridgeExchangeRateInfo.tsx
@@ -21,6 +21,8 @@ import { useBridgeState } from '@/slices/bridge/hooks'
 
 import { useAppSelector } from '@/store/hooks'
 
+const MAX_ARB_REBATE_PER_ADDRESS = 2000
+
 export const BridgeExchangeRateInfo = () => {
   return (
     <div className="py-3.5 px-1 space-y-3 text-sm md:px-6 tracking-wide">
@@ -72,7 +74,15 @@ const RouteEligibility = () => {
   const { isRouteEligible, isActiveRouteEligible, rebate } =
     useStipEligibility()
 
-  if (!isRouteEligible || !rebate) {
+  const { parsedCumulativeRewards } = useAppSelector(
+    (state) => state.feeAndRebate
+  )
+
+  if (
+    !isRouteEligible ||
+    !rebate ||
+    Number(parsedCumulativeRewards) > MAX_ARB_REBATE_PER_ADDRESS
+  ) {
     return (
       <div className="flex justify-between">
         <div className="flex-grow" />
@@ -124,7 +134,15 @@ const RebateText = () => {
 const Rebate = () => {
   const { isRouteEligible, rebate } = useStipEligibility()
 
-  if (!isRouteEligible || !rebate) {
+  const { parsedCumulativeRewards } = useAppSelector(
+    (state) => state.feeAndRebate
+  )
+
+  if (
+    !isRouteEligible ||
+    !rebate ||
+    Number(parsedCumulativeRewards) > MAX_ARB_REBATE_PER_ADDRESS
+  ) {
     return null
   }