+++ packages/synapse-interface/contexts/BackgroundListenerProvider.tsx
@@ -2,6 +2,7 @@ import { createContext } from 'react'
 import { useApplicationListener } from '@/utils/hooks/useApplicationListener'
 import { useBridgeListener } from '@/utils/hooks/useBridgeListener'
 import { usePortfolioListener } from '@/utils/hooks/usePortfolioListener'
+import { useRiskEvent } from '@/utils/hooks/useRiskEvent'
 import { useTransactionListener } from '@/utils/hooks/useTransactionListener'
 import { use_TransactionsListener } from '@/utils/hooks/use_TransactionsListener'
 
@@ -14,6 +15,7 @@ export const BackgroundListenerProvider = ({ children }) => {
   useTransactionListener()
   use_TransactionsListener()
   useBridgeListener()
+  useRiskEvent()
 
   return (
     <BackgroundListenerContext.Provider value={null}>