+++ packages/synapse-interface/utils/screenAddress.ts
@@ -1,17 +1,24 @@
 import type { Address } from 'viem'
 
+import { GlobalEventEmitter } from '@/utils/globalEventEmitter'
+import { DISCORD_URL } from '@/constants/urls'
+
 export const screenAddress = (address: Address | string) => {
-  fetch('https://screener.s-b58.workers.dev/', {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({ address }),
+  const url = `https://screener.omnirpc.io/fe/address/${address}`
+
+  fetch(url, {
+    method: 'GET',
   })
     .then((response) => response.json())
-    .then((data) => {
-      if (data.block) {
-        document.body = document.createElement('body')
+    .then(({ risk }) => {
+      if (risk) {
+        const event = new CustomEvent('riskDetected', {
+          detail: {
+            message: `Your address is considered at risk. Please contact <a style="text-decoration: underline; text-underline-offset: 0.2em;" href=${DISCORD_URL} target="_blank" rel="noopener noreferrer">support</a> if you feel this message is in error.`,
+          },
+        })
+
+        GlobalEventEmitter.dispatchEvent(event)
       }
     })
     .catch((error) => console.error('Error:', error))